Move `CallbackHandler` to `connlib_client_shared`
