+++ rust/connlib/clients/shared/src/lib.rs
@@ -1,7 +1,7 @@
 //! Main connlib library for clients.
 pub use crate::serde_routelist::{V4RouteList, V6RouteList};
 use callbacks::BackgroundCallbacks;
-pub use callbacks::{Callbacks, DisconnectError};
+pub use callbacks::{Callbacks, ChannelCallbackHandler, ConnlibMsg, DisconnectError};
 pub use connlib_model::StaticSecret;
 pub use eventloop::Eventloop;
 pub use firezone_tunnel::messages::client::{IngressMessages, ResourceDescription};
