+++ rust/headless-client/src/ipc_service.rs
@@ -1,7 +1,7 @@
-use crate::{CallbackHandler, ConnlibMsg};
 use anyhow::{Context as _, Result, bail};
 use atomicwrites::{AtomicFile, OverwriteBehavior};
 use clap::Parser;
+use connlib_client_shared::ConnlibMsg;
 use connlib_model::ResourceView;
 use firezone_bin_shared::{
     DnsControlMethod, DnsController, TOKEN_ENV_KEY, TunDeviceManager, device_id, device_info,
@@ -594,8 +594,7 @@ impl<'a> Handler<'a> {
         .context("Failed to create `LoginUrl`")?;
 
         self.last_connlib_start_instant = Some(Instant::now());
-        let (cb_tx, cb_rx) = mpsc::channel(1_000);
-        let callbacks = CallbackHandler { cb_tx };
+        let (callbacks, cb_rx) = connlib_client_shared::ChannelCallbackHandler::new();
 
         // Synchronous DNS resolution here
         let portal = PhoenixChannel::disconnected(
